int -> double

+++ src/main/java/xyz/samiki/zollsystem/Inventorys/Inventorys.java
@@ -6,7 +6,7 @@ import xyz.samiki.zollsystem.controllers.ItemController;
 
 public class Inventorys {
 
-    public static Inventory getConfirmation(int price) {
+    public static Inventory getConfirmation(double price) {
         Inventory inv = InventoryController.inventoryPrepare(36,"§4Kostenpflichtig §8Passieren?");
         inv.setItem(11, ItemController.getItem(Material.LIME_WOOL,"§aJa"));
         inv.setItem(15, ItemController.getItem(Material.RED_WOOL,"§cNein"));
